Poject setup : Code refactor
